Move treemap rendering modules into src/treemap subdirectory
