+++ src/treemap/rasteriseSvg.test.ts
