+++ src/treemap/svgRasteriser.ts
