+++ src/treemap/svgRasteriserError.test.ts
