+++ src/treemap/treemapModel.ts
